Created Activity_Checker.sh
This bash script defines a custom signal handler that terminates the program when inputted. The script also monitors which users log into and out of the machine on which the code is run. When detected, the username and time stamp is reported on the terminal. Checks occur in intervals of 10 seconds.