Created minor4svr.c - pt. 1 of project LINK
This C project has two parts: a client program and a server program. Each program is made to connect. When successful, "PING" and "PONG" messages appear on the server side showing full-duplex communication. Meanwhile, the client side reports the RTTs or Round-trip time (RTT) in milliseconds per network request. When the connection ends, the client displays a summary of packet info such as the amount lost, amount transmitted, and max/min RTTs.